create generate pdf routes
